The Shopping App
When the user goes to the mall to shop, she/he has a basket with a size of 2000 grams. The user adds products to the basket. During shopping, the user can check the items that are in the basket already. If an item that the user selects cannot be added to the basket because of exceeding the basket capacity, the program should print a warning message. Shopping is finished when the user enters finish shopping, or the basket is full. This projects covers the topics of: Strings, Arrays, File I/O, Abstract Data Types, Generics, Bags.